+++ lib/to_source/emitter/alias.rb
@@ -1,5 +1,28 @@
 module ToSource
   class Emitter
+    class VAlias < self
+
+      handle(Rubinius::AST::VAlias)
+
+    private
+
+      delegate(:to, :from)
+
+      # Perform dispatch
+      #
+      # @return [undefined]
+      #
+      # @api private
+      #
+      def dispatch
+        emit('alias ')
+        emit(to)
+        space
+        emit(from)
+      end
+
+    end
+
     # Emiter for alias node
     class Alias < self